requirements-dev.txt
# Coverage
coverage==5.0.3

# DebugToolbar
django-debug-toolbar==2.2

-r requirements.txt